requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
arrow==1.3.0
attrs==23.2.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
blinker==1.7.0
bottle==0.12.25
bottle-websocket==0.2.9
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.2.0
cookiecutter==2.6.0
distro==1.9.0
dlib==19.24.2
dnspython==2.6.1
Eel==0.16.0
enum34==1.1.10
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.13.3
Flask==3.0.0
flet==0.20.2
flet-core==0.20.2
flet-runtime==0.20.2
frozenlist==1.4.1
fsspec==2024.3.1
future==0.18.3
gevent==23.9.1
gevent-websocket==0.10.1
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
greenlet==3.0.2
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
hugchat==0.4.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
keyboard==0.13.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==1.26.3
oauthlib==3.2.2
ollama==0.1.6
openai==0.28.0
opencv-python==4.9.0.80
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
Pillow==10.1.0
pip==24.0
playsound==1.2.2
plyer==2.1.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.17.2
pymongo==4.6.2
PyMsgBox==1.0.9
pynput==1.7.6
pyparsing==3.1.1
pyperclip==1.8.2
pypiwin32==223
pypng==0.20220715.0
PyQt6==6.6.1
PyQt6-Qt6==6.6.2
PyQt6-sip==13.6.0
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.8.2
python-slugify==8.0.4
pyttsx3==2.90
pytweening==1.0.7
pytz==2024.1
pywhatkit==5.4
pywin32==306
pywinauto==0.6.8
PyYAML==6.0.1
qrcode==7.4.2
repath==0.9.0
replicate==0.24.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
rsa==4.9
selenium==4.16.0
setuptools==69.0.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.10.1
speedtest==0.0.1
sympy==1.12
text-unidecode==1.3
torch==2.2.2
tqdm==4.66.2
trio==0.23.2
trio-websocket==0.11.1
ttkthemes==3.2.2
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0
watchdog==3.0.0
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==3.0.1
whichcraft==0.6.1
wikipedia==1.4.0
wsproto==1.2.0
yarl==1.9.4
zope.event==5.0
zope.interface==6.1
